Add ability to specify the pressure offset at initialization
Fixes #19.

The SCD30 only allows you to do this at startup, but the pressure offset
is hard-coded in the library. Add an optional pressure offset to
Adafruit_SCD30::begin.

This is an API-breaking change, since objects that override
Adafruit_SCD30::_init will have the wrong signature. This is done
intentionally; any objects that override _init need to take into account
the new argument to Adafruit_SCD30::begin.

Also get rid of the unused decl for getAmbiendPressure.